review-app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
databricks-sdk>=0.30.0
pydantic>=2.0.0
pypdf>=4.0.0
python-multipart>=0.0.9
requests>=2.31.0
openpyxl>=3.1.0
